Вывод node links

+++ List/src/struct.h
@@ -6,8 +6,8 @@ static const node_val_type POISON = 0xDED;
 
 typedef struct node
 {
-    size_t next = 0;          //
-    size_t prev = 0;          //
+    int next = 0;          //
+    int prev = 0;          //
     node_val_type value  = 0; // 
 
 }mode;
@@ -16,9 +16,9 @@ typedef struct list
 {
     node*  nodes_arr        = nullptr; //
 
-    size_t head_node        = 0;       //
-    size_t tail_node        = 0;
-    size_t free_node        = 0;       //
+    int head_node           = 0;       //
+    int tail_node           = 0;
+    int free_node           = 0;       //
 
     size_t cur_num_of_nodes = 0;       //
     size_t max_num_of_nodes = 0;       //
